Reorganize file upload saga modules.
