+++ client/src/saga/uploadBallotManifestSaga.ts
@@ -1,6 +1,6 @@
 import { takeEvery } from 'redux-saga/effects';
 
-import notice from '../../notice';
+import notice from '../notice';
 
 
 function* importBallotManifestOk(action: any): IterableIterator<any> {
